Remove "right now" phrase
The phrase "right now" indicates whether this is a permanent thing or something that will change over time. A better way to explain what will change is to change it if it changes or provide a brief description, such as explaining whether this is something on the to-do list or left alone.