updated README.md
Affected URL(s)
https://bun-docs.vercel.app/#/?id=configuring-with-environment-variables

Description of the problem
Under the Credits section
On line no. 2, it is mentioned as based off of whereas based on would be the correct choice of a word representing it as official documentation of Bun.
I hope this would help to make the Bun Documentation more precise to the users.