add --save argument to install
When bun encounters a long opt argument which it doesn't understand, it
immediately stops parsing all remaining arguments (see
src/deps/zig-clap/clap/streaming.zig:96). This leads to some weird
behavior:

1. We currently don't define `--save` as a possible argument to `bun
   install`. This meant that `bun install --save X` would be incorrectly
   parsed as `bun install`.
2. On the other hand, `bun install X --save` would be parsed as `bun
   install X`. But that would be redirected to `bun add` and parsed as
   `bun add X`, which works fine.

To fix this, define `--save` as a possible argument for `bun install`.
This in turn requires some refactoring because we used to use the
presence of `--save` in the argument parser definition to determine if
the subcommand should default to modifying package.json or not. Instead
we introduce a Subcommand enum and clean that logic up, along with some
similar subcommand-specific logic.

A future diff may fix the long opt behavior, assuming I'm not missing a
good reason for its existence. :)

Fixes: https://github.com/oven-sh/bun/issues/3203