fix: don't remove content-encoding header from header table
It seams the 'content-encoding' header was being removed(if present),
to run code conditionally, that had the unwanted side effect of the header
not being present in the list returned to the caller.
This instead uses a bool to keep track of the state regarding the state whether
the header was prcoessed or not.
Maybe this has some other side effects but tests(fetch.test.ts) seams to
not have any failures.