fix(bun install): Handle vercel and github tarball path dependencies
package.json dependency entries like:

```
"when": "https://github.com/cujojs/when/tarball/1.0.2"
```

and

```
"@vercel/turbopack-ecmascript-runtime": "https://gitpkg-fork.vercel.sh/vercel/turbo/crates/turbopack-ecmascript-runtime/js?turbopack-230922.2"
```

are both links to tarballs. In the github case, this is a URL pattern, and in the vercel case, this seems to be everything served from their server.