fix syscall error handling
The error handling between calling zig syscalls and using libc is
different. Within zigs implementation, and therefore it's clone within
bun, the return value of a syscall is of type usize. Libc implemenations
use a signed integer namely c_int. As soon as libc is linked every
response code of a syscall is interpreted via the libc errno function,
but not every syscall is made through the libc interface. Only a few
syscalls on aarch64 are used from libc every other one is called through
zig itself. This could lead to errors beeing misinterpreted as success,
cause its strictly positive and therefore not an error for the libc
implementation. If an open syscall internally fails but is interpreted
as a success the following write syscall would operate on a wrong file
descriptor resulting in error bad file descriptor (EBADF).

Fixes #4723

Signed-off-by: Tobias Kohlbau <tobias@kohlbau.de>