fix: don't inline process.env calls in runtime
This caused issues with reassigning or deleting env vars as the inlining into a static
value meant that the changes would not be reflected at the runtime, while maybe it would be
somehow possible to do static analyisis whether a value is changed that is massively out of scope.
This disables the inlining of environment variables at transpile time.

Fixes: https://github.com/oven-sh/bun/issues/6762
Fixes: https://github.com/oven-sh/bun/issues/4383