fix: correctly call abort handler when request is aborted
We where not protecting the signal js object on the request when it was requested,
leading to the garbage collection of it, preventing handlers like `on("abort")` and `signal.onabort = ...`
from being executed on the JS side.
This fixes that by storing the js value and unprotecting it when required.

Fixes: https://github.com/oven-sh/bun/issues/4517