test: Fix segfault on timeout while awaiting
Fixes: #8069

The crash was because JSValue.zero was returned to JSC by way of Expect.getValue, without a JSC exception being raised to prevent the value from being dereferenced. The code was trying to use throwTerminationException for this purpose, but that implementation wasn't actually throwing an exception.

Future work: figure out why the exception only gets logged sometimes, because it would be nice to always get the source dump and stack trace on timeout.