fix(posix): use fcntl to check if a fd is nonblocking (#10080)
The FileReader is responsible for deciding what sort of file descriptor
has been passed to it. On MacOS, when stdin is provided to it the
FileReader assumes that it is nonblocking even if the descriptor is not
as it passes the pollable and not a tty check, this then results in
hangs as the reader expects `EAGAIN`.

To fix this we now check the file descriptor to see if it is nonblocking
rather than using assumptions.