Restore support for dynamic function names in tracebacks
Support for this was removed in 68e6fe00a4be7857f474cb3aa80e0d876499e3f8
during the migration to bun.String, but that completely breaks using the
`::bunternal::` name to hide stack frames, which is utilized by multiple
places in the existing codebase. It also matches Node's own behavior
when printing stack traces.

(Note that this does not touch the CallSite- / Error.stack-related
code, which is separate from this.)