Fix AES-GCM encryption of empty messages
AES-GCM is not block based, unlike AES-CBC, so we're not required to pad
the ciphertext. Instead, a tag is written at the end to authenticate the
plaintext after decryption. `EVP_EncryptFinal_Ex` within this context is
used as a way to signal that we've concluded writing our data and should
proceed to compute the authentication tag. In the case of a zero
message, this logic would essentially get triggered via the earlier call
to `EVP_EncryptUpdate`, resulting in an error when `EVP_EncryptFinal_Ex`
is then called because it is not idempotent. To fix, we simply check
that we have some plaintext to write before writing it.