build: use generic x86_64 microarchitecture levels
Currently the Intel CPU families Nehalem and Haswell are used to denote
"baseline" and "modern" builds. This creates a bit of a misnomer in the
former, as Nehalem is not actually the baseline microarchitecture level
for x86_64, but x86_64-v2. Likewise, Haswell is not the modern x86_64-v3
target but a superset of it with instructions that may or may not be
present on equivalent x86_64-v3 (non-Haswell) processors. See
https://en.wikipedia.org/wiki/X86-64#Microarchitecture_levels for a
general overview as well as the differences between these targets in
zig's stdlib here https://ziglang.org/documentation/master/std/#std.Target.x86.cpu