test: add failing test for #13696
It seems that the http module does not stream the initial few chunks of
a streaming http response in real-time. This has caused issues when
using bun with testcontainers.

While I am not qualified to locate and resolve the root cause of this
issue, I can provide a breaking test, in the hopes it will help somebody
else locate and fix the issue.