Move Bun-specific V8 classes into namespace v8::shim
The v8 namespace was confusingly polluted with both classes that exist
in the actual V8 API, as well as new classes to support Bun's specific
implementation of that API. This commit moves the latter into the
namespace v8::shim. It also makes a duplicate of classes with fields in
the v8::shim namespace. For those classes, the v8:: version has no
fields, no JSC superclass, and only implements the actual V8 API
functions, while the v8::shim:: version has the inheritance and
functions necessary to be allocated as a JSC object and implements
helper functions to support the V8 APIs.