Update `@types/node` dependency version
When I upgraded my project to use TypeScript@5.6 I ran into the issue
described in [this SO post](https://stackoverflow.com/a/78790944/223931).

The solution is to use a newer version of `@types/node` which has been
updated to deal with the changes in the latest TypeScript.

In my project I've introduced an explicit dependency on `@types/node@20.16.15`
(the newest version currently available in the `^20.0.0` range) since the
fix doesn't exist in `20.12.14` (the newest version available in the `~20.12.8`
range which was currently defined) which seems to solve the problem. But I
thought it would be helpful to others if the change was made here so they
didn't need to debug the problem.

DISCLAIMER: I don't know how these types are used and if there is any chance
I've introduced other problems with this change...