kit + server component improvements
you can use --server-components with bun build, and you get a slightly
more correct output. this output is not yet perfect or usable. but i am
committing this as a checkpoint as the next step is a large design
change for how `kit.DevServer` bundles and re-bundles files

we are removing the old `Framework` struct, and replacing it with a new
one with a similar purpose: configuration options targeted towards
framework developers. it tells the bundler how to glue the framework to
the application, mostly in regards to react server components

introduce `bundle_v2.AstBuilder`, which is used for parallel generation
of virtual modules. previously, this used a string builder combined with
the javascript parser. this is silly and will not be fun to use for the
3-5 file kinds that the bundler will have to generate. this builder
makes it very easy

introduce `kit.DevServer.IncrementalGraph`, which will keep track of
changes between bundle tasks. however, this code is not ripe and will
be iterated on extensively. i don't think it's approach is sound yet