begin hot reloader iteration 2
there are two `IncrementalGraph`s. one for the client and one for the
server. the plan is all routes share these two graphs, de-duplicating
cross-route work. more explanation is left in documentation comments.

more server components bugs are fixed, but not all. it is in a weird
state where some files are tree-shaken away, as well as bogus output
like `const Client = Client;`