fix(bun:test): Filtering of tests defined inside .each
Normally, 'describe' blocks are evaluated even when their name doens't
match the given test pattern, because the tests inside the block could
still match. However, with .each, this was instead skipping the
evaluation altogether, which made it near-impossible to filter the tests
inside.

This updates the code to match the regular filtering code, which will
always run the blocks inside and just change the tag to mark skipped
tests.