fix: parse JSX namespace identifiers that have numbers in them
Curiously, namespaced attribute in JSX allow numbers in the name.
However, the lexer code for namespaced attribute names does not
take numbers after the : character, so this results in an error
that should not exist; this commit allows for numbers to exist
in these namespaced attribute names.

This also adds some tests for this new behavior.