Avoid unnecessary chmod when creating symlinks
When running bun install, bun does a chmod on binaries before creating
symlinks. See this strace snippet:

    symlink("../workerd/bin/workerd", "/app/node_modules/.bin/workerd") = 0
    chmod("/app/node_modules/workerd/bin/workerd", 0777) = 0 <9.825437>

When run in a container, chmod triggers a file to be copied from the overlay
file system. The workerd binary above is 96 MB, which takes 9 or 10 seconds to
copy and there are 3 of them. This is causing bun install to take 30+ seconds.

I'd like to run bun install and then chmod 0777 the files in the Dockerfile.
Then at runtime, if we run bun install again, a stat should be able to reveal
that the chmod is unnecessary. The stat won't trigger an expensive copy-up like
chmod would.

So this commit just has bun check if the file already has correct 0777
permissions before calling chmod.