Display symlink target path when failing to open one
This helps with debugging as it makes it clear to the user the target is ENOENT and not the actual path they are passing to `Bun.file`. This is borne out of my own confusion at seeing the path exists with `ls` and seeing the entry in my VS Code Explorer pane, but not realizing the symlink target has been deleted or the path was even a symlink to begin with.

Fixes #17557