Rename fromJSHostValue -> fromJSHostCall and use a CatchScope
This function is changed to perform the function call itself rather than
taking the return value. The reason is that, if the function we're
calling has a ThrowScope, JSC wants us to check for its exception using
a ThrowScope or CatchScope that was declared *before* the callee's scope
was declared. Under the old pattern, fromJSHostValue could only do
anything *after* the callee had returned. Now it can surround the callee
in a CatchScope. To enable this it also takes a global object and a
SourceLocation.