Added a devcontainer configuration
In this initial version, the container environment succesfully builds the project,
though many of the tests fail. This seems to be a larger/unrelated problem?

I've need to switch from compressing debug symbols with zlib instead of zstd, as
in the few base image distro's I've tried, I couldn't find one that has support
for this compiled into clang, while also providing a libstdc++ that doesn't clash
with the clang version (regarding constexpr std::string).