Fix `File` initialization in `initEmpty`
Previously this function initialized files in an invalid state, where
`flags.kind` and `content` didn't match. This was changed to always
initialize as `unknown`, but this broke logic that relied on seeing a
`FileKind` of `.css`. This commit adds an additional field to the tagged
`content` union to handle non-root CSS files that don't have asset IDs.