Hide private fields in typeInfo (#22228)
https://github.com/oven-sh/zig/compare/ebe0cdac3170b0462a39caff4d1d7252b608e98d..e0b7c318f318196c5f81fdf3423816a7b5bb3112

It hides it in typeInfo even in the current file, unlike private
declarations which are only hidden in other files.

This allows you to formatted print a type with private fields, but the
private fields are not shown. The alternative would be to allow
accessing private fields through `@field()` but that looked like it was
going to be more complicated (need to add an argument to
structFieldVal/structFieldPtr which are called by fieldVal/fieldPtr
which have 36 callsites)

---------

Co-authored-by: taylor.fish <contact@taylor.fish>