refactor(completion): allow parsing of package.json with `,` and spaces in values
The majority of previous changes was uneccesary complication due the
using of `sed` to filter the completion candidates, i.e. when user types
`bun run script-name <Tab>` and the `script-name` is the existing key
under package.json `"scripts"`, then script suggestinon are filtred out.

This change elminate the usage of `sed` for the purpose of filtering at
this particular case. The bash's loops and case constructs are used
instead.

The few usages of `compgen` have also eliminated, so the splitting array
on spaces is avoidet altogether (for scripts only). So now the script
names with spaces inside is also allowed.