fix(completion): fix and supress shellcheck's SC2207
Introduces the proper reading values not under programmer's control into
array. In places where the value splitted into array is constant or
literal the warning is supressed maintaining the simplicity and
readability.

It is worth to note, that `compgen` outputs `\n` when result of
glob-expansion of its argument (with `-X` flag set) is empty. This
causes an empty element in the completion array. Additional check has
been added to avoid such condition.

See https://www.shellcheck.net/wiki/SC2207 for more info.
Also https://stackoverflow.com/a/45201229/21575560 lists the wrong and
offers the right way to read character-delemited string into array.

In this commit the locally set globstar feature has also been refactored
to use the trap making restoring the state of option manually
uneccessary preventing future bugs with forgeting to restore at the end
of the function.