fix(completion): use bash's syntax syntax for variable usage site consistenly
In many cases where variable's content is prefixed or suffixed the error can arise such that `"$MY_VARiable"` will be substituted with nothing. This can be dangerous in matching cases and patterns. Furthemore, Bash eliminates the emptyness, so unset variable in argument position is treated as omited argument, so the next argument if any (and not empty) steps into the role of the previous one. This is somewhat trappy situaton.