fix(completion): introduce escaping of bash specials for files and scripts
With this change an user can use suggestion as is without quoting.
Without it, an user are offered the two posibilities:

1. Start with `"` or `'`, then triger completion with `<Tab>`.
2. Trigger completion, then move cursor back, quoting.

The former does not work in bash. The latter is tedious.

The complete list of special characters can be found in bash-wiki.

bash-wiki: https://mywiki.wooledge.org/BashGuide/SpecialCharacters