fix(completion): change filtering method for file names
`globstar` comes from the upstream code. This option allows `**` inside globs matching zero or more subdirectories, but this double-star was and is not actually in use anywhere, hence it safe to remove it.

Important: `compgen -X <glob>` uses globs for flitering, so using regexps with it is the error. Therefore, the `=~` check was introduced.