fix(completion): correct building of long & short options lists
The previous implementation erroneously matched currently completed word
against `-`, so the branch for long options, `--`,  was never executed.

Even replacing first `^-` with `^--` might not impove anything, since
the first condition is still incorrectly filters candidates even using
the empty current word, hence the third branch is added in this commit.

Additionaly, it simplifies the first parameter of the function and
supplies corresponding simplified arguments into it.