refactor(completion): eliminate extra script filtering loop
The previous version of the script used two pass algorithm: the first pass to populate `COMPREPLY`, the second one to filter script back from it, if previous word is a script name. This approach is essantially erroneous, sinse the filtering condition can be used at the first pass.

To accomodate to the fix `COMPREPLY` should not be mutated inside the loop, the if loop exits without "return", `COMPREPLY` can be populated with script names.

This commit also refactors the conditionals for displaying completion in case of unmatched first subcommand/option/flag and the empty current word, since this part of code is also a call site of the function parsing package.json file. It uses flagging variable to detect the "previous word is script name" condition, as the previos version of the script did.