chore: suggest correct lockfile upgrade command
The flag `--frozen-lockfile` is a yarn flag, not an npm flag. 

If the goal is to only do a lockfile upgrade, not an install, `npm --package-lock-only` is the flag.

But if the intent is to get `npm ci` or `yarn install --frozen-lockfile` type divergence protection, neither of those can write to the lockfile anyways so they couldn't perform the version upgrade.