Fix inotify for directory watches
There are two core changes to how we use inotify itself:
1. Register IN_MODIFY interest when watching a directory. This gives
   access to `change` events on files inside the watched directory.
   Because we want access to these events, this also removes the blanket
   `rename` casting in the directory register path.
2. Treat the `IN_CREATE` bit as `rename`. This is to preserve existing
   behavior from when we casted all events in the directory registration
   path to `rename`.

This also changes the core logic for consecutive duplicate event
detection. The previous directory watch implementation would emit events
if `NOT (current_event.type != last_event.type AND current_event.hash !=
last_event.hash)`.  Concretely, if you emitted `rename` then `change` on the same
path, the event check would pass, but the hash one would mean the
expression returns `False`. That means that the `NOT` block would
evaluate to `True`, and we'd miss the event.

This logic was fine previously since we didn't register IN_MODIFY and
thus would cast all events on a watched directory to a `rename`, so the
event types would always miss the check. Now that we register with
IN_MODIFY,  we have to pick up `change` events as well.

Lastly, this properly updates the last event's hash. It seems that got
lost in a refactor (see 006575a0f10).